requirements.txt
pyopengl==3.1.0
pillow==5.4.1
numpy
typing
enum
